+++ app/tests/e2e/test_connections.py
@@ -2,9 +2,9 @@ import time
 
 import pytest
 from assertpy import assert_that
-from app.error.cloud_api_error import CloudApiException
 
-from app.tests.util.member_personas import BobAliceConnect
+from app.error.cloud_api_error import CloudApiException
+from app.tests.util.ecosystem_connections import BobAliceConnect
 from app.tests.util.webhooks import check_webhook_state
 from app.util.rich_async_client import RichAsyncClient
