+++ app/tests/e2e/test_fixtures.py
@@ -1,9 +1,7 @@
-from typing import Any
-
 import pytest
 from aries_cloudcontroller import AcaPyClient
-from app.admin.tenants.models import CreateTenantResponse
 
+from app.admin.tenants.models import CreateTenantResponse
 from app.dependencies import acapy_auth, acapy_auth_verified
 from app.generic.definitions import (CreateCredentialDefinition, CreateSchema,
                                      CredentialSchema,
@@ -11,7 +9,7 @@ from app.generic.definitions import (CreateCredentialDefinition, CreateSchema,
                                      create_schema)
 from app.generic.issuer.issuer import router
 from app.listener import Listener
-from app.tests.util.ecosystem_personas import FaberAliceConnect
+from app.tests.util.ecosystem_connections import FaberAliceConnect
 from app.tests.util.ledger import create_public_did, has_public_did
 from app.tests.util.string import random_version
 from app.tests.util.trust_registry import register_issuer
