+++ app/tests/e2e/test_webhook.py
@@ -2,7 +2,7 @@ import pytest
 
 from app.error.cloud_api_error import CloudApiException
 from app.generic.webhooks import router
-from app.tests.util.member_personas import BobAliceConnect
+from app.tests.util.ecosystem_connections import BobAliceConnect
 from app.util.rich_async_client import RichAsyncClient
 from shared_models import Connection
